refactor: review dto 패키지 분리
