+++ src/main/java/ssu/eatssu/domain/review/dto/request/ReviewUpdateRequest.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain.review.dto;
+package ssu.eatssu.domain.review.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Max;
