+++ src/main/java/ssu/eatssu/domain/review/dto/response/MenuIdNameDto.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain.review.dto;
+package ssu.eatssu.domain.review.dto.response;
 
 public record MenuIdNameDto(Long id,
                             String name) {
