+++ src/main/java/ssu/eatssu/domain/review/dto/response/MenuReviewResponse.java
@@ -1,4 +1,7 @@
-package ssu.eatssu.domain.review.dto;
+package ssu.eatssu.domain.review.dto.response;
+
+import ssu.eatssu.domain.review.dto.RatingAverages;
+import ssu.eatssu.domain.review.dto.ReviewRatingCount;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
