+++ src/main/java/ssu/eatssu/domain/review/presentation/ReviewControllerV2.java
@@ -18,21 +18,21 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import ssu.eatssu.domain.auth.security.CustomUserDetails;
 import ssu.eatssu.domain.restaurant.entity.Restaurant;
-import ssu.eatssu.domain.review.dto.CreateMealReviewRequest;
-import ssu.eatssu.domain.review.dto.CreateMenuReviewRequestV2;
-import ssu.eatssu.domain.review.dto.MealReviewResponse;
-import ssu.eatssu.domain.review.dto.MealReviewsV2Response;
-import ssu.eatssu.domain.review.dto.MenuReviewsV2Response;
-import ssu.eatssu.domain.review.dto.RestaurantReviewResponse;
-import ssu.eatssu.domain.review.dto.ReviewDetail;
-import ssu.eatssu.domain.review.dto.ReviewTranslationResponse;
-import ssu.eatssu.domain.review.dto.UpdateMealReviewRequest;
-import ssu.eatssu.domain.review.dto.ValidMenuForViewResponse;
+import ssu.eatssu.domain.review.dto.request.CreateMealReviewRequest;
+import ssu.eatssu.domain.review.dto.request.CreateMenuReviewRequestV2;
+import ssu.eatssu.domain.review.dto.response.MealReviewResponse;
+import ssu.eatssu.domain.review.dto.response.MealReviewsV2Response;
+import ssu.eatssu.domain.review.dto.response.MenuReviewsV2Response;
+import ssu.eatssu.domain.review.dto.response.RestaurantReviewResponse;
+import ssu.eatssu.domain.review.dto.response.ReviewDetail;
+import ssu.eatssu.domain.review.dto.response.ReviewTranslationResponse;
+import ssu.eatssu.domain.review.dto.request.UpdateMealReviewRequest;
+import ssu.eatssu.domain.review.dto.response.ValidMenuForViewResponse;
 import ssu.eatssu.domain.review.presentation.docs.ReviewControllerV2Docs;
 import ssu.eatssu.domain.review.service.ReviewServiceV2;
 import ssu.eatssu.domain.review.service.ReviewTranslationService;
 import ssu.eatssu.domain.slice.dto.SliceResponse;
-import ssu.eatssu.domain.user.dto.MyMealReviewResponse;
+import ssu.eatssu.domain.user.dto.response.MyMealReviewResponse;
 import ssu.eatssu.domain.user.entity.Language;
 import ssu.eatssu.global.handler.response.BaseResponse;
