+++ src/main/java/ssu/eatssu/domain/review/presentation/docs/ReviewControllerDocs.java
@@ -14,13 +14,13 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.web.multipart.MultipartFile;
 import ssu.eatssu.domain.auth.security.CustomUserDetails;
 import ssu.eatssu.domain.menu.entity.constants.MenuType;
-import ssu.eatssu.domain.review.dto.MealReviewsResponse;
-import ssu.eatssu.domain.review.dto.MenuReviewResponse;
-import ssu.eatssu.domain.review.dto.ReviewCreateRequest;
-import ssu.eatssu.domain.review.dto.ReviewDetailV1;
-import ssu.eatssu.domain.review.dto.ReviewUpdateRequest;
-import ssu.eatssu.domain.review.dto.SavedReviewImage;
-import ssu.eatssu.domain.review.dto.UploadReviewRequest;
+import ssu.eatssu.domain.review.dto.response.MealReviewsResponse;
+import ssu.eatssu.domain.review.dto.response.MenuReviewResponse;
+import ssu.eatssu.domain.review.dto.request.ReviewCreateRequest;
+import ssu.eatssu.domain.review.dto.response.ReviewDetailV1;
+import ssu.eatssu.domain.review.dto.request.ReviewUpdateRequest;
+import ssu.eatssu.domain.review.dto.response.SavedReviewImage;
+import ssu.eatssu.domain.review.dto.request.UploadReviewRequest;
 import ssu.eatssu.domain.slice.dto.SliceResponse;
 import ssu.eatssu.global.handler.response.BaseResponse;
