+++ src/main/java/ssu/eatssu/domain/review/service/ReviewService.java
@@ -11,14 +11,14 @@ import ssu.eatssu.domain.menu.entity.Menu;
 import ssu.eatssu.domain.menu.persistence.MealRepository;
 import ssu.eatssu.domain.menu.persistence.MenuRepository;
 import ssu.eatssu.domain.rating.entity.RatingCalculator;
-import ssu.eatssu.domain.review.dto.MealReviewsResponse;
-import ssu.eatssu.domain.review.dto.MenuReviewResponse;
+import ssu.eatssu.domain.review.dto.response.MealReviewsResponse;
+import ssu.eatssu.domain.review.dto.response.MenuReviewResponse;
 import ssu.eatssu.domain.review.dto.RatingAverages;
-import ssu.eatssu.domain.review.dto.ReviewCreateRequest;
+import ssu.eatssu.domain.review.dto.request.ReviewCreateRequest;
 import ssu.eatssu.domain.review.dto.ReviewRatingCount;
-import ssu.eatssu.domain.review.dto.ReviewUpdateRequest;
-import ssu.eatssu.domain.review.dto.SavedReviewImage;
-import ssu.eatssu.domain.review.dto.UploadReviewRequest;
+import ssu.eatssu.domain.review.dto.request.ReviewUpdateRequest;
+import ssu.eatssu.domain.review.dto.response.SavedReviewImage;
+import ssu.eatssu.domain.review.dto.request.UploadReviewRequest;
 import ssu.eatssu.domain.review.entity.Review;
 import ssu.eatssu.domain.review.entity.ReviewImage;
 import ssu.eatssu.domain.review.repository.ReviewImageRepository;
