+++ src/main/java/ssu/eatssu/domain/review/service/ReviewServiceV2.java
@@ -17,25 +17,25 @@ import ssu.eatssu.domain.menu.persistence.MenuRepository;
 import ssu.eatssu.domain.menu.service.MealRatingService;
 import ssu.eatssu.domain.rating.entity.Ratings;
 import ssu.eatssu.domain.restaurant.entity.Restaurant;
-import ssu.eatssu.domain.review.dto.CreateMealReviewRequest;
-import ssu.eatssu.domain.review.dto.CreateMenuReviewRequestV2;
-import ssu.eatssu.domain.review.dto.MealReviewResponse;
-import ssu.eatssu.domain.review.dto.MealReviewsV2Response;
-import ssu.eatssu.domain.review.dto.MenuIdNameDto;
-import ssu.eatssu.domain.review.dto.MenuLikeRequest;
-import ssu.eatssu.domain.review.dto.MenuReviewsV2Response;
+import ssu.eatssu.domain.review.dto.request.CreateMealReviewRequest;
+import ssu.eatssu.domain.review.dto.request.CreateMenuReviewRequestV2;
+import ssu.eatssu.domain.review.dto.response.MealReviewResponse;
+import ssu.eatssu.domain.review.dto.response.MealReviewsV2Response;
+import ssu.eatssu.domain.review.dto.response.MenuIdNameDto;
+import ssu.eatssu.domain.review.dto.request.MenuLikeRequest;
+import ssu.eatssu.domain.review.dto.response.MenuReviewsV2Response;
 import ssu.eatssu.domain.review.dto.RatingAverages;
-import ssu.eatssu.domain.review.dto.RestaurantReviewResponse;
-import ssu.eatssu.domain.review.dto.ReviewDetail;
+import ssu.eatssu.domain.review.dto.response.RestaurantReviewResponse;
+import ssu.eatssu.domain.review.dto.response.ReviewDetail;
 import ssu.eatssu.domain.review.dto.ReviewRatingCount;
-import ssu.eatssu.domain.review.dto.UpdateMealReviewRequest;
-import ssu.eatssu.domain.review.dto.ValidMenuForViewResponse;
+import ssu.eatssu.domain.review.dto.request.UpdateMealReviewRequest;
+import ssu.eatssu.domain.review.dto.response.ValidMenuForViewResponse;
 import ssu.eatssu.domain.review.entity.Review;
 import ssu.eatssu.domain.review.repository.ReviewRepository;
 import ssu.eatssu.domain.review.repository.ReviewTranslationRepository;
 import ssu.eatssu.domain.review.utils.MenuFilterUtil;
 import ssu.eatssu.domain.slice.dto.SliceResponse;
-import ssu.eatssu.domain.user.dto.MyMealReviewResponse;
+import ssu.eatssu.domain.user.dto.response.MyMealReviewResponse;
 import ssu.eatssu.domain.user.entity.User;
 import ssu.eatssu.domain.user.repository.UserRepository;
 import ssu.eatssu.global.handler.response.BaseException;
