+++ src/main/java/ssu/eatssu/domain/slice/service/SliceService.java
@@ -10,12 +10,12 @@ import ssu.eatssu.domain.menu.entity.Menu;
 import ssu.eatssu.domain.menu.entity.constants.MenuType;
 import ssu.eatssu.domain.menu.persistence.MealRepository;
 import ssu.eatssu.domain.menu.persistence.MenuRepository;
-import ssu.eatssu.domain.review.dto.ReviewDetail;
-import ssu.eatssu.domain.review.dto.ReviewDetailV1;
+import ssu.eatssu.domain.review.dto.response.ReviewDetail;
+import ssu.eatssu.domain.review.dto.response.ReviewDetailV1;
 import ssu.eatssu.domain.review.entity.Review;
 import ssu.eatssu.domain.review.repository.ReviewRepository;
 import ssu.eatssu.domain.slice.dto.SliceResponse;
-import ssu.eatssu.domain.user.dto.MyReviewDetail;
+import ssu.eatssu.domain.user.dto.response.MyReviewDetail;
 import ssu.eatssu.domain.user.entity.User;
 import ssu.eatssu.domain.user.repository.UserRepository;
 import ssu.eatssu.global.handler.response.BaseException;
